First Commit
Created a context with two devices: one CPU and one GPU. The devices reside in a cl_device_id pointer called devices. The first element is the CPU and the second the GPU. Built the matrix_mult.c program for the context, but have yet to write the actual kernel.