Working on the Kernel side
Printing matrix inputs and multiplied matrix output within the kernel (after a global memory barrier) shows that the computation is correct. The problem I'm having now is that it seems like matrices A and B in the host program get corrupted after copying them to a memory buffer for sending to the kernel. Similarly, although the output is correct within the kernel reading the buffer after computation gives me some crazy values.